Remove required field polyfill
This polyfill prevents the submission of forms (both `remote` and
standard) if there are blank required inputs. This behavior is
implemented natively by most modern browsers, the sole holdout being
Safari.

When the behavior is handled natively by the browser, the user is shown
appropriate error messages. When the behavior falls through the
Polyfill, there is no notification to the user as to why their form now
fails to submit. To the user, the form simply appears broken. The source
of this behavior is difficult for developers to track down.

jquery-ujs has a [longstanding issue][1]on this polyfill. With the move
to a new gem for ujs in Rails 5.1 we can make this breaking change and
remove the incomplete polyfill.

[1]:https://github.com/rails/jquery-ujs/issues/319